some minor cleanup, template reformat for styling. filtering by pathogenicity threshold on visualizable variant

+++ lirical-core/src/main/java/org/monarchinitiative/lirical/core/output/OutputOptions.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 public record OutputOptions(LrThreshold lrThreshold,
                             MinDiagnosisCount minDiagnosisCount,
                             float pathogenicityThreshold,
-                            Double frequencyThreshold,
+                            float frequencyThreshold,
                             Path outputDirectory,
                             String prefix,
                             Iterable<OutputFormat> outputFormats) {
